backend/requirements.txt
fastapi
uvicorn
pydantic
google-genai
torch
tokenizers
numpy
python-dotenv
sqlalchemy
passlib[bcrypt]
pyjwt
python-multipart
psycopg2-binary